Fix issue where Strict parameter is not present
Since the Strict parameter is set to omitifempty, if it is set to
false it will not be present in a request.

The idiomatic way of resolving this in golang is to make it a bool
pointer. This way, you can instantiate a bool and set it to false to
ensure that it is present.